+++ libromi/test/mock/mock_scriptengine.h
@@ -5,5 +5,5 @@
 class MockScriptEngine : public romi::ScriptEngine<romi::Rover>
 {
         MOCK_METHOD(int, get_next_event, (), (override));
-        MOCK_METHOD(void, execute_script, (romi::Rover& target, int index), (override));
+        MOCK_METHOD(void, execute_script, (romi::Rover& target, size_t index), (override));
 };
